requirements.txt
langchain
openai
pinecone-client
python-dotenv
PyPDF2
cassio
tiktoken
astrapy
flask
flask_session
python-docx
exception